Support odd keys like BSharp
With our previous logic, we treated BSharp Major and
C Major the same. We should reject BSharp Major.

Ultimately, this came down to our accidental counters
using BaseNote, which treats BSharp and C the same.
The right answer here is to realize that if the tonic of
your key has an accidental, your key cannot possibly
 have 0 accidentals. We are thus intercepting the
counters in these cases and enforcing that 0
accidentals means invalid.